+++ src/gfx/rgbacolor.h
@@ -43,35 +43,35 @@ namespace Gfx
 class Gfx::RgbaColor
 {
 private:
-  double itsData[4];
+  float itsData[4];
 
 public:
-  RgbaColor(double v = 1.0)
+  RgbaColor(float v = 1.0)
     { itsData[0] = itsData[1] = itsData[2] = itsData[3] = v; }
 
-  RgbaColor(double r_, double g_, double b_, double a_)
+  RgbaColor(float r_, float g_, float b_, float a_)
     { set(r_, g_, b_, a_); }
 
-        double& r()       { return itsData[0]; }
-  const double& r() const { return itsData[0]; }
+        float& r()       { return itsData[0]; }
+  const float& r() const { return itsData[0]; }
 
-        double& g()       { return itsData[1]; }
-  const double& g() const { return itsData[1]; }
+        float& g()       { return itsData[1]; }
+  const float& g() const { return itsData[1]; }
 
-        double& b()       { return itsData[2]; }
-  const double& b() const { return itsData[2]; }
+        float& b()       { return itsData[2]; }
+  const float& b() const { return itsData[2]; }
 
-        double& a()       { return itsData[3]; }
-  const double& a() const { return itsData[3]; }
+        float& a()       { return itsData[3]; }
+  const float& a() const { return itsData[3]; }
 
-  void get(double& r_, double& g_, double& b_, double& a_) const
+  void get(float& r_, float& g_, float& b_, float& a_) const
     { r_ = r(); g_ = g(); b_ = b(); a_ = a(); }
 
-  void set(double r_, double g_, double b_, double a_)
+  void set(float r_, float g_, float b_, float a_)
     { r() = r_; g() = g_; b() = b_; a() = a_; }
 
-        double* data()       { return &itsData[0]; }
-  const double* data() const { return &itsData[0]; }
+        float* data()       { return &itsData[0]; }
+  const float* data() const { return &itsData[0]; }
 };
 
 static const char __attribute__((used)) vcid_groovx_gfx_rgbacolor_h_utc20050626084023[] = "$Id$ $HeadURL$";
